Refactor test lifecycle to use summary table setup
